+++ packages/rsi-vanilla/src/interpretation-player.js
@@ -56,6 +56,7 @@ const getRoomName = () => {
  * @returns {Object} { sdkKey:string, positionMenu:string, isBoxShadow:string, isPlayerControlled:string, roomName:string  }
  */
 const getConfig = () =>{
+  console.info(component.dataset,'component.dataset');
   let {
      sdkKey:apiKey,
      positionMenu,