+++ packages/rsi-vanilla/update-version.js
@@ -17,7 +17,8 @@ const path = require('path')
 var packageJson = require('./package.json'); 
 const currentVersion = packageJson.version
 const myBucket = 'akkadu-assets'
-const filePath = path.resolve('./packages/rsi-vanilla/dist/index.min.js')
+const filePathInterpretationPlayer = path.resolve('./packages/rsi-vanilla/dist/interpretation-player.min.js')
+const filePathInterpretationManager = path.resolve('./packages/rsi-vanilla/dist/interpretation-manager.min.js')
 
 let tmpVersionFile = require(path.resolve('./packages/rsi-vanilla/tmp-version.json'));
 let tmpVersion;
@@ -33,7 +34,7 @@ if(tmpVersionFile && tmpVersionFile.tmpVersion){
  if(tmpVersion === currentVersion){
   console.info("No need to update rsi-vanilla on AWS akkadu-assets bucket. The version didn't change");
   return
-}  
+}   
 /**
  * @description We write a new tmpVersion file
  */
@@ -41,19 +42,19 @@ if(tmpVersionFile && tmpVersionFile.tmpVersion){
   const newTmpVersion =  { tmpVersion: currentVersion }
   const toJSON  = JSON.stringify(newTmpVersion, null, "\t")
   fs.writeFile(`./packages/rsi-vanilla/tmp-version.json`, toJSON, function(err, result) {
-    if(err) console.log('error', err);
+    if(err) console.error('error', err);
    });
 }
 
 /**
  * @description To update the vanillaJs Version documentation with the new version.
  */
- const concatMarkdownDoc = () =>{
+ const concatMarkdownDoc = (fileName) =>{
   fs.readFile('./rsi-documentation/docs/vanilla-js/versions.md', 'utf8', (err, doc) => {
-    const newVersion = `- Version ${currentVersion}: https://akkadu-assets.s3.amazonaws.com/akkadu-rsi-widget/rsi-vanilla/${currentVersion}/index.min.js`
+    const newVersion = `- Version ${currentVersion}: https://akkadu-assets.s3.amazonaws.com/akkadu-rsi-widget/rsi-vanilla/${currentVersion}/${fileName}.min.js`
     const newDoc = `${doc} \n ${newVersion}`
     fs.writeFile(`./rsi-documentation/docs/vanilla-js/versions.md`, newDoc, function(err, result) {
-      if(err) console.log('error when updating the doc with the new version', err);
+      if(err) console.error('error when updating the doc with the new version', err);
      });
   });
 }
@@ -62,23 +63,32 @@ if(tmpVersionFile && tmpVersionFile.tmpVersion){
 console.info('🚧🚧🚧');
 console.info('Updating the version on our bucket akkadu-rsi-widget/rsi-vanilla/, it will take a few minutes ...');
 console.info('🚧🚧🚧');
-console.log('currentVersion',currentVersion);
+console.info('currentVersion',currentVersion);
 
-AWS.config.update({ accessKeyId: process.env.AWS_ACCESS_KEY, secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY, });
-  fs.readFile(filePath, function (err, data) {
+const updateAWS = (file, fileName) =>{
+  fs.readFile(file, function (err, data) {
     if (err) { throw err; }
     var base64data = new Buffer(data, 'binary');
     var s3 = new AWS.S3();
     s3.putObject({
       Bucket: myBucket,
-      Key: `akkadu-rsi-widget/rsi-vanilla/${currentVersion}/index.min.js`,
+      Key: `akkadu-rsi-widget/rsi-vanilla/${currentVersion}/${fileName}.min.js`,
       Body: base64data,
       ACL: 'public-read'
     },function (resp) {
-      console.log(`Successfully uploaded the new version of rsi-widget on AWS, version: ${currentVersion}.`);
+      console.info(`Successfully uploaded the new version of rsi-widget on AWS, version: ${currentVersion}.`);
       writeTmpVersion();
-      concatMarkdownDoc()
+      concatMarkdownDoc(fileName)
     });
 });
+}
+
+
+AWS.config.update({ accessKeyId: process.env.AWS_ACCESS_KEY, secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY, });
+
+
+updateAWS(filePathInterpretationPlayer, 'interpretation-player')
+updateAWS(filePathInterpretationManager, 'interpretation-manager')
+