+++ packages/rsi-interpretation-manager/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-manager@0.1.0...@akkadu/rsi-interpretation-manager@0.1.1) (2021-07-02)
+
+**Note:** Version bump only for package @akkadu/rsi-interpretation-manager
+
+
+
+
+
+# [0.1.0](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-manager@0.0.17...@akkadu/rsi-interpretation-manager@0.1.0) (2021-07-02)
+
+
+### Features
+
+* added event language & interpretation language section ([1041808](https://github.com/Akkadu/rsi-api-widgets/commit/1041808523f911d7c901c6317a3695f7885b5b36))
+
+
+
+
+
 ## [0.0.17](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-manager@0.0.16...@akkadu/rsi-interpretation-manager@0.0.17) (2021-07-01)
 
 **Note:** Version bump only for package @akkadu/rsi-interpretation-manager