+++ packages/rsi-vanilla/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.10](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.9...@akkadu/rsi-vanilla@1.0.10) (2021-07-02)
+
+**Note:** Version bump only for package @akkadu/rsi-vanilla
+
+
+
+
+
+## [1.0.9](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.7...@akkadu/rsi-vanilla@1.0.9) (2021-07-02)
+
+**Note:** Version bump only for package @akkadu/rsi-vanilla
+
+
+
+
+
 ## [1.0.7](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.6...@akkadu/rsi-vanilla@1.0.7) (2021-07-01)
 
 **Note:** Version bump only for package @akkadu/rsi-vanilla